fix: various (small) fixes for registry operations
Pass auth to the registry frontend so that it correctly authenticates to
the registry on uploads/checks.

Refactor the code in the registry frontend to use digest instead of tag
when redirecting request for additional security.

In the schematic registry storage, push a manifest on top of raw blob to
make sure it doesn't get garbage collected as orphaned. We never use the
manifest though.

Signed-off-by: Andrey Smirnov <andrey.smirnov@siderolabs.com>