feat: use OCI layout when passing images to the imager
Fixes #39

This fixes for all image pulls to be part of the Image Factory itself,
Talos `imager` now never pulls on its own.

Installer image (base) is passed as an OCI layout to the `imager`, and
same for extension images (for consistency).

Virtual `schematic` extension is passed as `.tar`, as it doesn't make a
difference actually.

Signed-off-by: Andrey Smirnov <andrey.smirnov@siderolabs.com>