fix: drop extractParams function
In v0.7.5 there is a bug - if the final page of the wizard is reloaded,
the 'internal server error' is raised.

This was due to the type mismatch on the `wizardFinal` return type and
`extractParams` expected input type. This is bad development habbit to
have two types which should be kept in sync, so instead just drop this
function completely, as we don't need it.

Signed-off-by: Andrey Smirnov <andrey.smirnov@siderolabs.com>