fix: set content-disposition on S3
S3 does not set proper Content-Disposition, and this header might be
ignored by e.g. browsers if set before redirect. We need to include it
when adding new object to the cache.

Additionally, if the object was previously cached, we need to ensure
that Content-Disposition header will be properly updated.

Signed-off-by: Mateusz Urbanek <mateusz.urbanek@siderolabs.com>